+++ fon/Excitation_to_Formant.c
@@ -1,6 +1,6 @@
 /* Excitation_to_Formant.c
  *
- * Copyright (C) 1992-2002 Paul Boersma
+ * Copyright (C) 1992-2008 Paul Boersma
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,6 +20,7 @@
 /*
  * pb 1995/08/23
  * pb 2002/07/16 GPL
+ * pb 2008/01/19 double
  */
 
 #include "Excitation_to_Formant.h"
@@ -28,15 +29,15 @@ Formant Excitation_to_Formant (Excitation me, int maxnFormants)
 {
 	Formant thee = NULL;
 	long i, nfreq = my nx, nform = 0;
-	float *p = my z [1];
+	double *p = my z [1];
 
 	thee = Formant_create (0, 1, 1, 1, 0.5, maxnFormants); if (! thee) return NULL;
 	thy frame [1]. formant = NUMstructvector (Formant_Formant, 1, maxnFormants);
 	for (i = 2; i < nfreq; i ++)
 		if (p [i] > p [i - 1] && p [i] >= p [i + 1])
 		{
-			float min3phon, left, right;
-			float firstDerivative = p [i+1] - p [i-1], secondDerivative = 2 * p [i] - p [i-1] - p [i+1];
+			double min3phon, left, right;
+			double firstDerivative = p [i+1] - p [i-1], secondDerivative = 2 * p [i] - p [i-1] - p [i+1];
 			long j;
 			Formant_Formant formant = & thy frame [1]. formant [++ nform];
 			formant -> frequency = Excitation_barkToHertz (
